PoC/DW_PoC(GCP_vs_NiFi)/GCP/Pub:Sub/Twitter ToyPJ/requirements.txt
python-decouple==3.4
tweepy==3.10.0